常识融合GCObject到Object

+++ Source/Engine/Types/EngineState.hpp
@@ -3,6 +3,6 @@
 struct EngineState
 {
     float FPS = 0.0f;
-    float DeltaTime = 0.0f;
+    double DeltaTime = 0.0;
 	Vec2<int> ScreenSize;
 };
